performance_agent/requirements.txt
openai
pydantic
python-dotenv
numpy
pandas
scikit-learn
pyBKT
torch
azure-cosmos>=4.7.0
fastapi>=0.116.0
uvicorn>=0.35.0
python-multipart>=0.0.20
